backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pymongo==4.6.0
motor==3.3.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
python-multipart==0.0.6
scikit-learn>=1.4.0
numpy==1.24.3
pandas==2.1.3
tensorflow==2.15.0
keras==2.15.0
nltk==3.8.1
requests==2.31.0
python-dotenv==1.0.0
APScheduler==3.10.4
pdfminer.six==20221105
python-docx==1.1.0